+++ zenoh/src/api/session.rs
@@ -85,6 +85,7 @@ use crate::{
                 PublicationBuilderDelete, PublicationBuilderPut, PublisherBuilder,
                 SessionDeleteBuilder, SessionPutBuilder,
             },
+            query::SessionGetBuilder,
             queryable::QueryableBuilder,
             session::OpenBuilder,
             subscriber::SubscriberBuilder,
@@ -95,10 +96,7 @@ use crate::{
         info::SessionInfo,
         key_expr::{KeyExpr, KeyExprInner},
         publisher::{Priority, PublisherState},
-        query::{
-            ConsolidationMode, QueryConsolidation, QueryState, QueryTarget, Reply,
-            SessionGetBuilder,
-        },
+        query::{ConsolidationMode, QueryConsolidation, QueryState, QueryTarget, Reply},
         queryable::{Query, QueryInner, QueryableState},
         sample::{DataInfo, DataInfoIntoSample, Locality, QoS, Sample, SampleKind},
         selector::Selector,
